Added show-picked-versions to buildout.cfg
Our build shouldn't change just because a new version becomes available on PyPi. This ensures that buildout prints all the unspecified versions that had to be automatically picked and were unspecified. These should then be added under [versions] section on buildout.cfg.
A side-effect is that it speeds up buildout builds since it doesn't have to check for the latest versions available.